Added useClaimScreen hook
