+++ app/src/hooks/useClaimScreen.ts
@@ -1,8 +1,11 @@
+import axios, { AxiosError } from "axios";
+import React from "react";
 import { useState } from "react";
 import { Alert } from "react-native";
 import { useClaimsStore } from "../context/ClaimsStore";
+import { UploadPGPKeyResponse, VerifyEmail } from "../types/claim";
 import { check18Plus } from "../utils/birthday-utils";
-import { claimsLocalStorage } from "../utils/local-storage";
+import { claimsLocalStorage, pgpLocalStorage } from "../utils/local-storage";
 
 type Hooks = {
   loading: boolean;
@@ -11,11 +14,14 @@ type Hooks = {
   selectedFileIds: string[];
   setLoading: (loading: boolean) => void;
   setSelectedFileIds: (selectedFileIds: string[]) => void;
+  verifyEmail: (email: string) => void;
 };
 const useClaimScreen = (): Hooks => {
   const { addClaim } = useClaimsStore();
   const [loading, setLoading] = useState<boolean>(false);
   const [selectedFileIds, setSelectedFileIds] = useState<string[]>([]);
+  const [, setIsVerifying] = React.useState<boolean>(false);
+  const [, setVerifyEmailRequest] = React.useState<VerifyEmail>();
 
   const save18Claim = async () => {
     setLoading(true);
@@ -72,10 +78,70 @@ const useClaimScreen = (): Hooks => {
     }
   };
 
+  const verifyEmail = async (email: string) => {
+    setIsVerifying(true);
+    const armoredKey = await pgpLocalStorage.get();
+    if (!armoredKey || armoredKey === null) {
+      Alert.alert(
+        `ERROR`,
+        `Please go to your profile and import your private key first!`,
+        [
+          {
+            text: "Ok",
+            onPress: () => console.log("invalid pgp"),
+            style: "cancel"
+          }
+        ]
+      );
+      return;
+    }
+    try {
+      const uploadResponse = await axios.post<UploadPGPKeyResponse>(
+        "https://keys.openpgp.org/vks/v1/upload",
+        armoredKey,
+        {
+          headers: {
+            "Content-Type": "application/json"
+          }
+        }
+      );
+      setVerifyEmailRequest({
+        token: uploadResponse.data.token,
+        addresses: [email]
+      });
+      try {
+        await axios.post("https://keys.openpgp.org/vks/v1/upload", armoredKey, {
+          headers: {
+            "Content-Type": "application/json"
+          }
+        });
+        Alert.alert(
+          `Email Sent`,
+          `Please check your email for instructions from keys.openpgp.org on how to verify your claim.`,
+          [
+            {
+              text: "OK",
+              onPress: () => console.log(""),
+              style: "cancel"
+            }
+          ]
+        );
+      } catch (error) {
+        const err = error as AxiosError;
+        console.error(err?.response?.data || error);
+      }
+    } catch (error) {
+      const err = error as AxiosError;
+      console.error(err?.response?.data || error);
+    }
+    setIsVerifying(false);
+  };
+
   return {
     loading,
     saveAndCheckBirthday,
     onSelectFile,
+    verifyEmail,
     selectedFileIds,
     setLoading,
     setSelectedFileIds
