+++ app/src/navigation/screens/ClaimScreen.tsx
@@ -18,12 +18,10 @@ import {
   ProfileStackNavigationRoute
 } from "../../types/navigation";
 import { getClaimFromType } from "../../utils/claim-utils";
-import { Claim, UploadPGPKeyResponse, VerifyEmail } from "../../types/claim";
+import { Claim } from "../../types/claim";
 import { FileList, Button } from "../../components";
 import { useClaimsStore } from "../../context/ClaimsStore";
 import { useDocumentStore } from "../../context/DocumentStore";
-import { pgpLocalStorage } from "../../utils/local-storage";
-import axios, { AxiosError } from "axios";
 import { getDocumentFromDocumentId } from "../../utils/document-utils";
 import BottomNavBarSpacer from "../../components/BottomNavBarSpacer";
 import useClaimScreen from "../../hooks/useClaimScreen";
@@ -44,14 +42,14 @@ const ClaimScreen: React.FC = () => {
   const [showDatePickerForFieldId, setShowDatePickerForFieldId] =
     React.useState<string>();
   const [isVerifying, setIsVerifying] = React.useState<boolean>(false);
-  const [, setVerifyEmailRequest] = React.useState<VerifyEmail>();
 
   const {
     loading,
-    setLoading,
     saveAndCheckBirthday,
     onSelectFile,
+    verifyEmail,
     selectedFileIds,
+    setLoading,
     setSelectedFileIds
   } = useClaimScreen();
 
@@ -84,65 +82,6 @@ const ClaimScreen: React.FC = () => {
       routes: [{ name: "Home" }]
     });
     setLoading(false);
-  };
-
-  const verifyEmail = async (email: string) => {
-    setIsVerifying(true);
-    const armoredKey = await pgpLocalStorage.get();
-    if (!armoredKey || armoredKey === null) {
-      Alert.alert(
-        `ERROR`,
-        `Please go to your profile and import your private key first!`,
-        [
-          {
-            text: "Ok",
-            onPress: () => console.log("invalid pgp"),
-            style: "cancel"
-          }
-        ]
-      );
-      return;
-    }
-    try {
-      const uploadResponse = await axios.post<UploadPGPKeyResponse>(
-        "https://keys.openpgp.org/vks/v1/upload",
-        armoredKey,
-        {
-          headers: {
-            "Content-Type": "application/json"
-          }
-        }
-      );
-      setVerifyEmailRequest({
-        token: uploadResponse.data.token,
-        addresses: [email]
-      });
-      try {
-        await axios.post("https://keys.openpgp.org/vks/v1/upload", armoredKey, {
-          headers: {
-            "Content-Type": "application/json"
-          }
-        });
-        Alert.alert(
-          `Email Sent`,
-          `Please check your email for instructions from keys.openpgp.org on how to verify your claim.`,
-          [
-            {
-              text: "OK",
-              onPress: () => console.log(""),
-              style: "cancel"
-            }
-          ]
-        );
-      } catch (error) {
-        const err = error as AxiosError;
-        console.error(err?.response?.data || error);
-      }
-    } catch (error) {
-      const err = error as AxiosError;
-      console.error(err?.response?.data || error);
-    }
-    setIsVerifying(false);
   };
 
   const documentList =
@@ -223,7 +162,6 @@ const ClaimScreen: React.FC = () => {
                     value={formState[field.id] === "true"}
                     onValueChange={(value) => {
                       onChange(value ? "true" : "false");
-                      console.log();
                     }}
                   />
                 </View>
